iReady pomodoro, task list, and motivational quote app
iReady is a three-in-one app. Why download three different apps when you can have just one with all the functionality. My focus revolves around the pomodoro technique, hence the pomodoro timer. I struggle to take breaks when I work so the pomodoro technique really works for me and others! Going along with a pomodoro timer, I thought it would be very convenient to have a task list a part of the app, because this app is already focused on making work more efficient. Lastly, to get people excited to get their work done for the day, I decided to add a motivation quote api, that sends users notifications with motivational qoutes in them. They can customize the pomodoro timer and quotes, allowing them to have more quotes, what time of day to receive them, or none at all.